fox toy material change - loading circle needs fixing
1. tried to make the player change the material on the fox toy with the gaze-script gazeontrain; the material change works but the loading circle is off and keeps loading - can you fix it, please?
also the second interaction is to make the newspaper to disappear after the player looks at it - tried to use the same script with changing its materail to invisible - it does not work - can you fix that too?
the last thing is the fade in at the beginning of the experience - intro scene - can you have a look and adjust it please?